requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.3
django-cors-headers==3.2.1
django-modelcluster==5.0.1
django-modeltranslation==0.14.4
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.8
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.2
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.8
wagtail-autocomplete==0.5
wagtailgeowidget==5.0.2
webencodings==0.5.1
Willow==1.3
wrapt==1.11.2
